+++ prisma/migrations/20260511020546_add_clients_table/migration.sql
@@ -1,0 +1,16 @@
+-- CreateTable
+CREATE TABLE "clients" (
+    "id" SERIAL NOT NULL,
+    "user_id" INTEGER NOT NULL,
+    "id_type" VARCHAR(5) NOT NULL,
+    "id_number" VARCHAR(20) NOT NULL,
+    "name" TEXT NOT NULL,
+    "email" TEXT NOT NULL,
+    "address" TEXT,
+    "created_at" TIMESTAMP(3) NOT NULL DEFAULT CURRENT_TIMESTAMP,
+
+    CONSTRAINT "clients_pkey" PRIMARY KEY ("id")
+);
+
+-- AddForeignKey
+ALTER TABLE "clients" ADD CONSTRAINT "clients_user_id_fkey" FOREIGN KEY ("user_id") REFERENCES "users"("id") ON DELETE CASCADE ON UPDATE CASCADE;